requirements_dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements_dev.in
#
-e git+https://github.com/OpenDataServices/lib-cove-web.git@v0.14.0#egg=libcoveweb  # via -r requirements.txt
alabaster==0.7.12         # via sphinx
attrs==19.3.0             # via -r requirements.txt, hypothesis, jsonschema, pytest
babel==2.8.0              # via sphinx
bleach==3.1.5             # via -r requirements.txt, libcove, libcoveocds
cached-property==1.5.1    # via -r requirements.txt, libcove, libcoveocds, libcoveweb
certifi==2020.4.5.1       # via -r requirements.txt, requests, sentry-sdk
chardet==3.0.4            # via -r requirements.txt, requests
click==7.1.2              # via pip-tools
commonmark==0.9.1         # via -r requirements.txt, libcove, libcoveocds
contextlib2==0.6.0.post1  # via -r requirements.txt, schema
coverage==4.5.4           # via -r requirements_dev.in, coveralls, hypothesis, pytest-cov
coveralls==2.0.0          # via -r requirements_dev.in
dealer==2.1.0             # via -r requirements.txt, libcoveweb
defusedxml==0.6.0         # via -r requirements.txt, odfpy
django-bootstrap3==12.1.0  # via -r requirements.txt, libcoveweb
django-debug-toolbar==2.2  # via -r requirements.txt, libcoveweb
django-environ==0.4.5     # via -r requirements.txt, libcoveweb
django==2.2.14            # via -r requirements.txt, django-debug-toolbar, libcove, libcoveocds, libcoveweb
docopt==0.6.2             # via coveralls
docutils==0.16            # via sphinx
et-xmlfile==1.0.1         # via -r requirements.txt, openpyxl
flake8==3.8.2             # via -r requirements_dev.in
flattentool==0.11.0       # via -r requirements.txt, libcove, libcoveocds
hypothesis==3.52.3        # via -r requirements_dev.in
idna==2.9                 # via -r requirements.txt, requests
imagesize==1.2.0          # via sphinx
importlib-metadata==1.6.0  # via -r requirements.txt, flake8, jsonschema, pluggy, pytest
jdcal==1.4.1              # via -r requirements.txt, openpyxl
jinja2==2.11.2            # via sphinx
json-merge-patch==0.2     # via -r requirements.txt, libcove, libcoveocds, libcoveweb
jsonref==0.2              # via -r requirements.txt, flattentool, libcove, libcoveocds, libcoveweb
jsonschema==3.2.0         # via -r requirements.txt, libcove, libcoveocds, libcoveweb
lepl==5.1.3               # via -r requirements.txt, rfc6266
libcove==0.17.0           # via -r requirements.txt, libcoveocds, libcoveweb
libcoveocds==0.7.4        # via -r requirements.txt
libsass==0.20.0           # via -r requirements_dev.in
lxml==4.5.1               # via -r requirements.txt, flattentool
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==8.3.0     # via pytest
odfpy==1.4.1              # via -r requirements.txt, flattentool
openpyxl==2.6.4           # via -r requirements.txt, flattentool, libcoveweb
packaging==20.4           # via -r requirements.txt, bleach, pytest, sphinx
pip-tools==5.2.0          # via -r requirements_dev.in
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pycodestyle==2.6.0        # via flake8
pyflakes==2.2.0           # via flake8
pygments==2.6.1           # via sphinx
pyparsing==2.4.7          # via -r requirements.txt, packaging
pyrsistent==0.16.0        # via -r requirements.txt, jsonschema
pytest-cov==2.9.0         # via -r requirements_dev.in
pytest-django==3.9.0      # via -r requirements_dev.in
pytest-localserver==0.5.0  # via -r requirements_dev.in
pytest==5.4.2             # via -r requirements_dev.in, pytest-cov, pytest-django
python-dateutil==2.8.1    # via -r requirements.txt, libcoveweb
python-slugify==1.2.6     # via transifex-client
pytz==2020.1              # via -r requirements.txt, babel, django, flattentool
rangedict==0.1.6          # via -r requirements.txt
requests==2.23.0          # via -r requirements.txt, coveralls, libcove, libcoveocds, libcoveweb, sphinx, transifex-client
rfc3987==1.3.8            # via -r requirements.txt, libcove, libcoveweb
rfc6266==0.0.4            # via -r requirements.txt, libcoveweb
schema==0.7.2             # via -r requirements.txt, flattentool
selenium==3.141.0         # via -r requirements_dev.in
sentry-sdk==0.14.4        # via -r requirements.txt, libcoveweb
six==1.15.0               # via -r requirements.txt, bleach, jsonschema, libsass, packaging, pip-tools, pyrsistent, python-dateutil, transifex-client
snowballstemmer==2.0.0    # via sphinx
sphinx-rtd-theme==0.4.3   # via -r requirements_dev.in
sphinx==3.0.4             # via -r requirements_dev.in, sphinx-rtd-theme
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
sqlparse==0.3.1           # via -r requirements.txt, django, django-debug-toolbar
strict-rfc3339==0.7       # via -r requirements.txt, libcove, libcoveweb
transifex-client==0.13.9  # via -r requirements_dev.in
unidecode==1.1.1          # via python-slugify
urllib3==1.25.9           # via -r requirements.txt, requests, selenium, sentry-sdk, transifex-client
wcwidth==0.1.9            # via pytest
webencodings==0.5.1       # via -r requirements.txt, bleach
werkzeug==1.0.1           # via pytest-localserver
xmltodict==0.12.0         # via -r requirements.txt, flattentool, libcoveweb
zipp==1.2.0               # via -r requirements.txt, importlib-metadata, libcoveweb

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools

# these were added when adapting for deploying on heroku
waitress==1.4.3
whitenoise==5.0.1
psycopg2-binary==2.8.5

# For S3 storage
django-storages==1.9.1
boto3==1.13.14
smart-open==2.0.0
